refactored classes into packages, refactored application class, upped pom
