+++ src/main/java/mpo/mansedan/Application.java
@@ -4,10 +4,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ManSedanApplication {
+public class Application {
 
     public static void main(String[] args) {
-	SpringApplication.run(ManSedanApplication.class, args);
+	SpringApplication.run(Application.class, args);
 //	CSVFileReader.oneByOneExample();
 //	CSVFileReader.oneByOnePOJO();
     }
