+++ src/main/java/mpo/mansedan/controller/HelloController.java
@@ -1,4 +1,4 @@
-package mpo.mansedan;
+package mpo.mansedan.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
