+++ src/test/java/mpo/mansedan/ApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ManSedanApplicationTests {
+class ApplicationTests {
 
 	@Test
 	void contextLoads() {
